+++ src/index.ts
@@ -1,3 +1,4 @@
 'use strict';
 
 export { Planner } from './planner';
+export { Contract } from './base-contract';
